[FEAT] - 컴포넌트 폴더 분리 / 라우터 생성 / Home.js 주석처리
