+++ frontend/package.json
@@ -9,6 +9,8 @@
     "eslint-config-react-app": "^7.0.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-hook-form": "^7.49.2",
+    "react-router-dom": "^6.21.1",
     "react-scripts": "5.0.1",
     "sass": "^1.69.5",
     "web-vitals": "^2.1.4"
