+++ frontend/src/App.js
@@ -1,17 +1,22 @@
-import "./App.css";
-import Shelter from "./components/pages/shelter/Shelter";
-import Home from "./components/pages/home/Home";
-import Shelter from './components/pages/shelter/Shelter';
-import SignUp from './components/pages/signup/SignUp';
+import './App.css';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Shelter from './pages/shelter/Shelter';
+import SignUp from './pages/signup/SignUp';
+import SignIn from './pages/signin/SignIn';
+import Home from './pages/home/Home';
 
 function App() {
   return (
-    <>
-      <Home />
-      <Shelter />
-      <br />
-      <SignUp />
-    </>
+    <BrowserRouter>
+      <main>
+        <Routes>
+          {/* <Route path="/" element={<Home />} /> */}
+          <Route path="/user/signup" element={<SignUp />} />
+          <Route path="/user/signin" element={<SignIn />} />
+          <Route path="/shelter" element={<Shelter />} />
+        </Routes>
+      </main>
+    </BrowserRouter>
   );
 }
