+++ frontend/src/pages/signin/SignIn.js
@@ -1,0 +1,5 @@
+import './signin.scss';
+
+export default function SignIn() {
+  return <></>;
+}
